Changes to truth event selection and minor code fixes
The truth event selection code was moved to a separate cpp file for better readability.
The code was changed to only set truth event indices when the event is a real VBS
event. If not the case, truth indices are set to -1. Added also code
for debugging truth event info.

Separate log files per job according to process ID.

Add output tag information to README.